pyproject.toml
[tool.poetry]
name = "fast_kinda_solid_api"
version = "0.1.0"
description = ""
authors = ["Brandon Hamric <brandonhamric@gmail.com>"]
readme = "README.md"
packages = [{include = "fast_kinda_solid_api"}]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = ">=0.101,<1.0"
sqlalchemy = ">=2.0.14,<2.1"
uvicorn = ">=0.14.0,<1.0"
alembic = ">=1.10,<1.14"
asyncpg = ">=0.28.0,<0.30"
greenlet = ">=3.0.2,<3.1"
pydantic = ">=2.0.1,<3.0.0"
pytz = ">=2023.2,<2025.0"
httpx = ">=0.25,<0.30"
structlog = ">=22.2.0,<25.0.0"
python-json-logger = ">=2.0.5,<2.1"
opentelemetry-api = ">=1.22,<2.0"
opentelemetry-sdk = ">=1.22,<2.0"
opentelemetry-exporter-otlp-proto-grpc = ">=1.22,<2.0"
opentelemetry-exporter-otlp-proto-http = ">=1.22,<2.0"
pytest-asyncio = ">=0.23,<1.0.0"
pydantic-settings = ">=2.0.3,<3.0"
opentelemetry-instrumentation-fastapi = ">=0.46b0,<1.0"
opentelemetry-instrumentation-logging = ">=0.46b0,<1.0"

[tool.poetry.group.dev.dependencies]
poetry = ">=1.8.2,<2.0"
pytest = ">=8.1.1,<9.0"
pre-commit = ">=3.7.0,<4.0"
flake8 = ">=7.0.0,<8.0"
black = ">=24.4.2,<25.0"
isort = ">=5.13.2,<6.0"
mypy = ">=1.10.0,<2.0"
trio = ">=0.25.0,<1.0"
colorama = ">=0.4.6,<1.0"
autoflake = ">=2.3.1,<3.0"
types-pytz = ">=2023.2,<2025.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
